Move shared test code into the bdcs crate
This way the package database setup code can be shared by both unit
tests and integration tests. Compiling test code as part of the released
public interface isn't ideal, but it works for now.
